Wrapping up the engine
In preparations to complete the engine, I created a set of APIs that
simplify operations for the engine and abstracts the messier inner
workings. If I stay on schedule, I will complete the engine within a
week. I introduced thread safety onto the top most abstraction layers of
the engine. Also, the current API lends itself for integration into the
needed mod editor, which should enable realtime editing of the game
(including realtime crashes as the user loads newly created but buggy
scripts).