Completed loading stages in Level
In the spirit of what I am trying to do with the engine, I added a
LayerSystem (will handle the background layers of the level), a
TimerSystem for registering timers, and a TriggerManager class for
handling map triggers on the map! I fixed a potential bug in my Python
embedding implementation class (I made sure it is thread-safe)! I also
finished the loading stages for the level. Basically, loading a level
starts by instantiating the Level class.