Game engine compilation errors fixed!
All of the compilation errors in the game engine are fixed! The engine
is compiled into a set of shared libraries (dlls in win, so and dylib in
linux and mac [in the future]). For the engine to be truly complete, I
still have to create Lua and Python wrappers of the target API. However,
for testing purposes we can say we have a full engine!