Updated SWIG interfaces (stable)
I updated the old SWIG interface to accomodate the big changes to the
engine since a year ago. I actually had to delete the interface and copy
its contents to two separate interfaces tailored for each language (Lua
and Python). The autogenerated cxx files compile in their current form.
The interfaces also allowed me to spot a couple of undefined methods in
the engine API!