Added SoundQueue
SoundQueue is the class that will manage those sounds not directly
managed by instances of Unit (aka not managed by the UnitManager).
Basically, it's my way of allowing scripts to spawn sound effects in the
game. As the name implies, the sound effects are added to a queue and
each sound is played one by one! The SoundQueue also manages the
background music. It does this by keeping a backbuffer and a front
buffer. Basically, the front buffer is a pointer to the instance of
sound_base that is playing in a music channel. When a new music track is
added, the front buffer is faded and deleted, the front buffer is set to
the pointer held by the back buffer, the back buffer is set to a new
instance of sound_base, and the front buffer is faded in (which means it
is automatically playing). The soundqueue has an interface for dictating
how long the fade effects will take!