Added VideoPlayer class
I added a VideoPlayer class to the engine. This means that the engine
will rely on the FFmpeg libraries for video decoding! The upside is that
we can now play videos, yay! I attached an updated version of the game
engine diagram. I also made a couple of minor updates to the game and
the sound_base classes. I made an important update to the SoundQueue
class that will allow it to accommodate the asynchronous feeding of
sound portions. Hopefully, the sound won't be noticeably out of
synchronization with the video frames. We'll know when we get to test
it.