Update to SDL_TTF and Addition of the SQLite test
Updated the SDL_TTF library to the newer version that fixes a
requirenment of the old SDL.dll. The new version relies only on
SDL2.dll.
Finished implementing the SQLite unit test. With it I got a couple of
bugs fixed in the SQLite interface. My implementation of SQLite works
now!