Merge remote-tracking branch 'origin/TheIllusiveMan' into TheIllusiveMan
Conflicts:
	.objs/Unit_tests/base_components_test.o
	.objs/button.o
	.objs/draw_base.o
	.objs/sound_base.o
	.objs/textbox.o
	.objs/trigger.o
	.objs/ui.o
	.objs/unit.o
	Eureka.cbp
	Eureka.depend
	Eureka.layout
	Eureka_build_log.html
	Unit_tests/base_components_test.cpp
	bin/engine/Core.dll
	bin/engine/libCore.a
	bin/engine/libCore.def
	bin/tests/Core.dll
	bin/tests/base_components_test.exe
	draw_base.cpp
	draw_base.h
	sound_base.cpp
	texture_store.cpp
	texture_store.h

Added TextureStore class

All draw_base instances now store textures in
a static instance of
TextureStore. The reason for this is to make sure
textures are loaded
once and then every object that needs the texture
gets a pointer to it.
This strategy is my way of minimizing duplication
of textures when
loading environment objects like floor tiles! Rotation
and flipping
effects are done at draw time, so it doesn't affect the
texture.
However, transparency affects the texture prior to drawing, so
all
visible copies are affected by this effect. Textures are stored in a
BST
using uint64_t as the key type. The key is computed by applying a
custom
crc64 implementation on the texture file.