Added a command line parser to the engine
Also, added a standard stream redirection system (cout, cin, cerr).
Updated the locale unit test to further showcase the command line parser
(it can now be used as a standalone utility on top of being a test :D)
Added the CMD_Processor driver as the unit test for the command line
parser!