Changes to the engine and updated the Data directory templates
A quick update of what I have done so far. I fixed a couple of major
bugs in the engine's load methods. I added a simple locale system to the
IO manager in the engine, so that game texts can be condensed into a
"universal" script that is reexpanded with the target language. The idea
is that, using a locale descriptor file with all the sentences
translated to language x, an universal script can be used to ask the
engine for the sentence equivalent of language y in language x. The
locale system does not care if the sentences are encoded in unicode or
ansi. I added the unit test for the locale system. Coincidentally, the
unit test can be used as an utility for creating a universal script from
a text sample file and using the language descriptor file for the text's
language! The utility can be used from the command line. I added a
custom command line argument parser library I created to the engine. As
a result, the engine can now process command lines (once these are
specified within the engine or from a script executing in the engine
:D).