Recompilation with w64-mingw
Because of incompatibilities between Qt's "mingw4.8_32bit" kit and
Code::Block's default GCC compiler, I had to recompile the windows
version with Qt's toolchain. What I learned is that you have to make
sure the engine and editor are compiled with the same version of
libstdc++ if you want to prevent a runtime error with the operator = in
std::string (the application crashes because of some bad linker
assembly).